+++ components/ReusableMediumComponents/ToggeableAlert.tsx
@@ -3,17 +3,18 @@
 import type { Dispatch, SetStateAction } from "react";
 import GeneralButton from "@components/ReusableSmallComponents/buttons/GeneralButton";
 
-export type ToggeableAlertProps = {
-  text: string;
-  toggleState: boolean | string;
-  setToggleState: Dispatch<SetStateAction<boolean | string>>;
-};
+export type ToggeableAlertProps<T extends boolean | string = boolean | string> =
+  {
+    text: string;
+    toggleState: T;
+    setToggleState: Dispatch<SetStateAction<T>>;
+  };
 
-export default function ToggeableAlert({
+export default function ToggeableAlert<T extends boolean | string>({
   text,
   setToggleState,
   toggleState,
-}: ToggeableAlertProps) {
+}: ToggeableAlertProps<T>) {
   return (
     <div className="p-4 bg-red-900 text-subtleWhite text-center rounded-2xl mb-2 max-w-[90%] mx-auto border-2 border-yellow-200">
       <p className="self-center">{text}</p>
@@ -21,7 +22,11 @@ export default function ToggeableAlert({
       <GeneralButton
         text="Close"
         className="mx-auto"
-        onClick={() => setToggleState(!toggleState)}
+        onClick={() =>
+          setToggleState(
+            (typeof toggleState === "boolean" ? false : "") as T,
+          )
+        }
       />
     </div>
   );