+++ types/modules.d.ts
@@ -1,0 +1,3 @@
+declare module "react-google-recaptcha";
+
+declare module "mongoose-unique-validator";